+++ frontend/components/sites/SiteBasicData.vue
@@ -49,7 +49,7 @@ SPDX-License-Identifier: EUPL-1.2
         <v-divider class="my-4" />
         <v-row>
           <v-col cols="12" :md="siteImagesShouldBeRendered ? 12 : 6">
-            <label>Usage</label>
+            <label>Land use</label>
             {{ value.siteUsageName | orDefault }}
             <a v-if="value.siteUsageUri" target="_blank" :href="value.siteUsageUri">
               <v-icon small>